Refactor: 랜덤 메뉴 선정 파일 분리
